Add kindle watermarker script and homemount
kindle watermarker will take the first arg and watermark a bunch of
kindle screensaver size images with it. Like the "slide to unlock"
default messages.

homemount is to sshfs a bunch of directories and is useful to
no-one but me and then only on my home network.

bin files do not get automatically installed yet.