Resolve vundle & git issue
Vundle had cloned bundles into a subdir of my repository and I had added
them to git without adding them as submodules. This meant after cloning
that the directories were there, but git didn't know where to get their
contents from.

The solution is to remove and git ignore the directories, add vundle as
a submodule for convenience, and just let vundle handle fetching the
bundles.